# FIIM_OpenAPI.yaml
openapi: 3.0.3
info:
  title: FIIM API
  description: |
    Fatigue Injury Index Monitoring (FIIM) REST API.
    
    This API provides endpoints for athlete monitoring, training load management,
    injury risk calculation, and multi-disciplinary team collaboration.
    
    All endpoints are scoped by organization (multi-tenant). Include the 
    `X-Organization-ID` header for every request, or use the user's default 
    organization from the JWT token.
    
    Base URL: `https://api.fiim.io/v1`
  version: 1.0.0
  contact:
    name: FIIM Engineering Team
    email: api@fiim.io
servers:
  - url: https://api.fiim.io/v1
    description: Production
  - url: https://api-staging.fiim.io/v1
    description: Staging
  - url: https://api-dev.fiim.io/v1
    description: Development
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /auth/login:
    post:
      summary: Email/password login
      tags: [Authentication]
      security: [] # Public endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/RateLimitError'
  /auth/logout:
    post:
      summary: Logout and revoke session
      tags: [Authentication]
      responses:
        '204':
          description: Logout successful
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /auth/refresh:
    post:
      summary: Refresh access token
      tags: [Authentication]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: New token pair issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /auth/mfa/setup:
    post:
      summary: Setup MFA (TOTP)
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [method]
              properties:
                method:
                  type: string
                  enum: [totp, sms]
      responses:
        '200':
          description: MFA setup initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  secret:
                    type: string
                  qr_code_url:
                    type: string
                  backup_codes:
                    type: array
                    items:
                      type: string
  /auth/mfa/verify:
    post:
      summary: Verify MFA code during login
      tags: [Authentication]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [temp_token, mfa_code]
              properties:
                temp_token:
                  type: string
                mfa_code:
                  type: string
      responses:
        '200':
          description: MFA verified, full token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'

  # ==========================================================================
  # USERS & RBAC
  # ==========================================================================
  /users:
    get:
      summary: List users in organization
      tags: [Users]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: role
          in: query
          schema:
            type: string
        - name: is_active
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Paginated user list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'

    post:
      summary: Create user
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/{user_id}:
    get:
      summary: Get user details
      tags: [Users]
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

    patch:
      summary: Update user
      tags: [Users]
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated

    delete:
      summary: Deactivate user (soft delete)
      tags: [Users]
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: User deactivated
...
  /athletes:
    get:
      summary: List athletes
      tags: [Athletes]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: team_id
          in: query
          schema:
            type: string
            format: uuid
        - name: squad_id
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [active, loaned, injured, suspended, retired, transferred]
        - name: search
          in: query
          schema:
            type: string
          description: Search by name or email
      responses:
        '200':
          description: Paginated athlete list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Athlete'

    post:
      summary: Create athlete
      tags: [Athletes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAthleteRequest'
      responses:
        '201':
          description: Athlete created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Athlete'
  /athletes/{athlete_id}:
    get:
      summary: Get athlete details
      tags: [Athletes]
      parameters:
        - name: athlete_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Athlete details with latest indices
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AthleteDetail'

    patch:
      summary: Update athlete
      tags: [Athletes]
      parameters:
        - name: athlete_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAthleteRequest'
      responses:
        '200':
          description: Athlete updated

    delete:
      summary: Deactivate athlete
      tags: [Athletes]
      parameters:
        - name: athlete_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Athlete deactivated
...
  /sessions:
    get:
      summary: List training sessions
      tags: [Training Load]
      parameters:
        - name: team_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
        - name: session_type
          in: query
          schema:
            type: string
            enum: [technical, tactical, strength, competition, recovery, rehab, travel]
      responses:
        '200':
          description: Session list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TrainingSession'

    post:
      summary: Create training session
      tags: [Training Load]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
      responses:
        '201':
          description: Session created
  /sessions/{session_id}:
    get:
      summary: Get session details
      tags: [Training Load]
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Session with load records

    patch:
      summary: Update session
      tags: [Training Load]
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSessionRequest'
      responses:
        '200':
          description: Session updated
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/login or /auth/refresh
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for external integrations (public API v1)

  # ==========================================================================
  # PARAMETERS
  # ==========================================================================
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
      description: Page number for pagination
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
      description: Items per page (max 100)

  # ==========================================================================
  # RESPONSES
  # ==========================================================================
  responses:
    UnauthorizedError:
      description: Authentication required or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    RateLimitError:
      description: Too many requests
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
        X-RateLimit-Reset:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Error:
      type: object
      required: [type, title, status, detail]
      properties:
        type:
          type: string
          example: https://api.fiim.io/errors/unauthorized
        title:
          type: string
          example: Unauthorized
        status:
          type: integer
          example: 401
        detail:
          type: string
          example: Invalid or expired access token
        instance:
          type: string
          example: /auth/login
        trace_id:
          type: string
          example: abc123def456
...
    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
        pagination:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            total_count:
              type: integer
            total_pages:
              type: integer
            has_next:
              type: boolean
            has_prev:
              type: boolean

    # ------------------------------------------------------------------------
    # Authentication
    # ------------------------------------------------------------------------
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 12
        organization_id:
          type: string
          format: uuid
          description: Optional if user belongs to multiple orgs
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          example: 14400
        user:
          $ref: '#/components/schemas/User'

    # ------------------------------------------------------------------------
    # Users
    # ------------------------------------------------------------------------
    User:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        email:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [administrator, coach, sport_scientist, physiotherapist, sports_doctor, athlete]
        organization_id:
          type: string
          format: uuid
        mfa_enabled:
          type: boolean
        last_login:
          type: string
          format: date-time
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
    CreateUserRequest:
      type: object
      required: [email, first_name, last_name, role]
      properties:
        email:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [administrator, coach, sport_scientist, physiotherapist, sports_doctor, athlete]
        team_id:
          type: string
          format: uuid
        send_invitation:
          type: boolean
          default: true
    UpdateUserRequest:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        phone:
          type: string
        is_active:
          type: boolean
        mfa_enabled:
          type: boolean
...
    Athlete:
      type: object
      properties:
        athlete_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        first_name:
          type: string
        last_name:
          type: string
        date_of_birth:
          type: string
          format: date
        gender:
          type: string
          enum: [male, female, mixed, other]
        nationality:
          type: string
        email:
          type: string
        phone:
          type: string
        position:
          type: string
        jersey_number:
          type: string
        height_cm:
          type: integer
        weight_kg:
          type: number
        target_sleep_hours:
          type: number
        team_id:
          type: string
          format: uuid
        team_name:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
    AthleteDetail:
      allOf:
        - $ref: '#/components/schemas/Athlete'
        - type: object
          properties:
            latest_indices:
              $ref: '#/components/schemas/DailyIndices'
            latest_wellness:
              $ref: '#/components/schemas/WellnessSurveyResponse'
            latest_load:
              $ref: '#/components/schemas/DailyLoadSummary'
            active_injuries:
              type: array
              items:
                $ref: '#/components/schemas/InjuryCase'
            consent_status:
              type: string
              enum: [granted, pending, withdrawn]
    CreateAthleteRequest:
      type: object
      required: [first_name, last_name, date_of_birth]
      properties:
        first_name:
          type: string
        last_name:
          type: string
        date_of_birth:
          type: string
          format: date
        gender:
          type: string
        nationality:
          type: string
        email:
          type: string
        phone:
          type: string
        position:
          type: string
        jersey_number:
          type: string
        height_cm:
          type: integer
        weight_kg:
          type: number
        target_sleep_hours:
          type: number
          default: 8.0
        team_id:
          type: string
          format: uuid
        emergency_contact_name:
          type: string
        emergency_contact_phone:
          type: string
        blood_type:
          type: string
        known_allergies:
          type: string
    UpdateAthleteRequest:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        position:
          type: string
        jersey_number:
          type: string
        height_cm:
          type: integer
        weight_kg:
          type: number
        target_sleep_hours:
          type: number
        is_active:
          type: boolean
...
    WellnessSurveyResponse:
      type: object
      properties:
        response_id:
          type: string
          format: uuid
        athlete_id:
          type: string
          format: uuid
        survey_date:
          type: string
          format: date
        submitted_at:
          type: string
          format: date-time
        responses:
          type: object
          additionalProperties: true
        calculated_wellness_index:
          type: number
        wellness_index_confidence:
          type: string
        is_editable:
          type: boolean
...
    TrainingSession:
      type: object
      properties:
        session_id:
          type: string
          format: uuid
        team_id:
          type: string
          format: uuid
        name:
          type: string
        session_type:
          type: string
        scheduled_date:
          type: string
          format: date
        scheduled_start:
          type: string
        scheduled_end:
          type: string
        location:
          type: string
        planned_rpe:
          type: number
        planned_duration:
          type: integer
        status:
          type: string
        load_records_count:
          type: integer
    CreateSessionRequest:
      type: object
      required: [team_id, name, scheduled_date]
      properties:
        team_id:
          type: string
          format: uuid
        name:
          type: string
        session_type:
          type: string
          default: technical
        scheduled_date:
          type: string
          format: date
        scheduled_start:
          type: string
        scheduled_end:
          type: string
        location:
          type: string
        planned_rpe:
          type: number
        planned_duration:
          type: integer
    UpdateSessionRequest:
      type: object
      properties:
        name:
          type: string
        scheduled_start:
          type: string
        scheduled_end:
          type: string
        actual_start:
          type: string
          format: date-time
        actual_end:
          type: string
          format: date-time
        status:
          type: string
...
    DailyLoadSummary:
      type: object
      properties:
        summary_id:
          type: string
          format: uuid
        summary_date:
          type: string
          format: date
        total_srpe:
          type: integer
        total_distance_m:
          type: integer
        total_hsr_m:
          type: integer
        total_sprint_m:
          type: integer
        total_player_load:
          type: integer
        session_count:
          type: integer
        rest_day:
          type: boolean
        data_completeness_pct:
          type: number
...
    DailyIndices:
      type: object
      properties:
        athlete_id:
          type: string
          format: uuid
        calculation_date:
          type: string
          format: date
        wellness_index:
          type: number
        fatigue_index:
          type: number
        recovery_index:
          type: number
        workload_index:
          type: number
        acwr_ratio:
          type: number
        readiness_index:
          type: number
        injury_risk_index:
          type: number
        risk_zone:
          type: string
          enum: [green, yellow, red, crimson]
        readiness_zone:
          type: string
          enum: [green, yellow, amber, red]
        components:
          type: object
...